performance speed-up for js badchar check
Due to two characteristics of PDF streams the new isJavascript() variant
became extremely slow. First of all, we want to allow Javascript with
*some* bad characters. E.g., a trademark or unicode apostrophe would
already mark a Javascript stream as not being Javascript - this was
identified and fixed by Fernando Dominguez for the 0.3.5 release.
Secondly, due to the nature of the above, some streams were incorrectly
let through. This was caused by "\x00" not being counted as an incorrect
character and therefore the 10-20% quota not being met. Combined these
two issues let to a major performance issue, as whole streams were being
processed and included in the output.
This commit fixes those performance issues and replaces the Python
character counting by a regex query that's slightly faster.